refactor: separate command handling
