+++ src/db/schema.ts
@@ -1,12 +1,13 @@
 import { integer, sqliteTable, text } from "drizzle-orm/sqlite-core";
 import { ulid } from "ulid";
 
-export const messages = sqliteTable("messages", {
+export const messagesTable = sqliteTable("messages", {
 	id: text("id")
 		.primaryKey()
 		.$defaultFn(() => {
 			return ulid();
 		}),
-	userId: integer("user_id"),
+	messageId: text("messageId"),
+	userId: text("user_id"),
 	content: text("content"),
 });
